requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.8.0+cu128
transformers>=4.35.0
datasets>=2.14.0
pyyaml>=6.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
